test(web): tighten SNES frontend flow coverage
Add upper-case and negative SNES support assertions, extract the shared minimal SNES ROM helper, and make the SNES Playwright flow less flaky.

Closes #2823
